requirements.txt
albumentations==1.3.1
awkward==2.6.8
clam==3.2.10
dgl==1.2a231122+cu118
einops==0.8.0
force_relative_import==0.0.2.post0
hydra-core==1.3.2
imageio==2.31.3
kornia==0.7.3
lifelines==0.29.0
loguru==0.7.2
lxml==5.3.0
matplotlib==3.7.3
numba==0.58.1
numpy==1.24.1
omegaconf==2.3.0
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
openslide_python==1.3.1
pandas==2.2.3
Pillow==9.3.0
Pillow==10.4.0
prettytable==3.11.0
pycocotools==2.0.8
pytorch_lightning==2.0.8
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.3.0
scipy==1.14.1
Shapely==1.8.5
skimage==0.0
tensorflow==2.17.0
tifffile==2023.9.26
tiffslide==2.2.0
timm==0.9.8
torch==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.1
